Add an android_system_sepolicy fork with a build fix
A version of a commit for Android 12 (non-L) related to a security fix,
that misses a 12L-specific part, was picked there. Now builds are
broken, and despite that being brought up in the comments on Lineage's
gerrit (albeit without -1 V and CR votes), it was unfortunately pushed
without any edits. The fork contains a fixup with that missing
12L-specific part.

It seems that, not only is basic build testing not being done for
LineageOS 19.1, it might be that little of any attention in general
is also being paid. :(